==> pages/7_Terms.py <==
import streamlit as st

st.set_page_config(page_title="Picanalitika | Term of Services", layout="wide")

hide_st_style = """
            <style>
            MainMenu {visibility: hidden;}
            footer {visibility: hidden;}
            div.vis-configuration-wrapper{display: block; width: 400px;}
            </style>
            """
st.markdown(hide_st_style, unsafe_allow_html=True)

st.title('Term of Services')
st.markdown("Welcome new, current, or future Picanalitika customer! Picanalitika's online intelligence platforms monitor billions of digital documents daily & provide market insight, press communication tools, and social media analysis to customers worldwide.")

st.markdown("Picanalitika's terms and conditions of use can be found below. This Agreement constitutes a contract between the Customer (as defined in the Order Confirmation), and Picanalitika. These Terms govern the use of (and access to) the Services and the Picanalitika site.")

st.markdown('By using Picanalitika services, accessing the Picanalitika site, or by allowing anyone else to do so, the Customer agrees to be bound by these terms and conditions.')

 
st.header('How the Picanalitika License works')
st.markdown("Picanalitika's flagship services monitor online traditional, social, and broadcast media. Predicated on search filters & Boolean logic, the Picanalitika platform yields News Link Search Results in the form of hyperlinks. These hyperlinks link to a Third Party Site on which the article is published.")

st.header('A note about Spam')
st.markdown("Most Picanalitika Services allow forwarding search results to others. Customer agrees not to abuse this capability or Spam anyone using the Picanalitika Services and understands that Picanalitika may suspend or terminate any customer that Picanalitika believes to be doing so."

"To the extent permitted by law, Customer agrees to indemnify and hold harmless Picanalitika, its subsidiaries, affiliates, employees and customers, from and against any and all liability, damages, and fines (including reasonable attorneys’ fees) related to Customer’s breach of this warranty."

"Simply Put: Customer and all Users, must use the Site and the Picanalitika Services, and any content accessed by the Services, in accordance with all applicable laws, rules and regulations.")
st.header('Who can use Picanalitika’s Services?')
st.markdown("Purchase of a Picanalitika subscription grants a Customer a non-exclusive and non-transferable right for Customer’s Authorised Users (the quantity is specified in the Order Confirmation) to use the Services. While Customer may use the Services for any business purpose, it must be Customer’s own, on Customer’s company’s behalf. (This does not include partners or affiliates.) All Authorised Users must be employees of Customer’s company, unless they are Customer’s contracted external consultants who may use the Services exclusively on Customer’s behalf."

"If Customer is a marketing or public relations agency using the Picanalitika Services on behalf of its client(s), it agrees that:"

"The agreement between Customer and Customer’s client(s) is as protective of Picanalitika’s rights as this Agreement."
"That agreement permits Customer to provide Picanalitika with Customer’s client(s)’ information, and act on their behalf, if applicable."
"Picanalitika is not a party to Customer’s agreement with its client(s) and Customer is solely responsible for delivery of agency services to its client."
"All payments due under this Agreement must be satisfied regardless of the status of their payments to Customer.")
st.header('What is needed to use the Services?')
st.markdown("Not much! In most cases all that’s needed is a web browser, an internet connection, and a password. Customer is responsible for having all the hardware, and internet access to use the Services. Customer is responsible for the activity on its account, as well as protecting its own password; please let Picanalitika know as soon as there is suspected unauthorised use of the account or password.")

st.header('Payment')
st.markdown("The price and payment terms of the Services appear on the Order Confirmation. Payment can be made by electronic wire transfer. Ask a Picanalitika representative for more information."

"Unless otherwise specified in the Order Confirmation, payment is due upfront in a single lump sum and is invoiced at the start of the subscription (first-time customers) or in advance of the renewal (current customers). Customer is responsible for paying any applicable taxes which aren’t included in the price and will appear on the invoice."

"All purchased news outlets become accessible to the user immediately upon logging into a Picanalitika account. Because Picanalitika services are fully available to the Customer from the start of the subscription all payments are non-refundable (unless otherwise specified here). If Customer cannot satisfy its payment obligations by the invoice due date, Picanalitika has the right to charge a late fee, Late Payment Interest and/or suspend or terminate access to the Services.")

st.header('Uptime')
st.markdown("Picanalitika guarantees service levels of 99.5% uptime throughout the Term, one of the most aggressive by SaaS standards. If there were ever a technical outage that makes the Services unavailable to Customer for more than 3 consecutive business days following Customers notice to Picanalitika, Picanalitika will issue Customer a credit equal to the pro-rata length of the outage (based on Customer’s annual payment to Picanalitika).")

st.header('Term & Termination')
st.markdown(
"Picanalitika will provide Services for the term indicated by start and stop dates on the Order Confirmation. If this Agreement is terminated, all rights of use expire."
"Should Customer breach this Agreement, Picanalitika has the right to suspend access to the Service immediately upon discovering the breach; Customer will have 30 days to cure that breach at which point Picanalitika has the right to terminate."

"Should Picanalitika materially breach this Agreement (and can’t cure within 30 days of Customer’s written notice,) Customer is entitled to a pro-rata refund of the Services not delivered.")

st.header('Renewals')
st.markdown("To make renewal as easy as possible for customers, this Agreement includes an automatic renewal. Unless Customer cancels its next term at least 60 days before the end of the term, this Agreement will automatically renew for the same subscription length at Picanalitika's then standard price."
"If Picanalitika does not receive Customer’s written cancellation at least 60 days before the end of the term, Customer agrees to the renewal.")
 
st.header('Intellectual Property')
st.markdown("Picanalitika is proud of its proprietary software and Services. Picanalitika is pleased to grant Customer rights of use to these Services, and those rights are strictly limited to the ones granted here in Agreement. Customer agrees not to sell, re-sell, reproduce, duplicate, copy, modify, or otherwise exploit any portion or derivative of the Site or its contents; nor to use any data mining, robots, or similar data gathering and extraction tools."

"All Services and all content on the Site (unless created by Customer or a third party) including software, code, trademarks or logos are owned by or licensed to Picanalitika, and Picanalitika reserves all rights not expressly granted in these Terms. Customer agrees not to use Picanalitika's name, trademark, logo, or Picanalitika generated content without Picanalitika's express written consent.")

st.header('Warranty & Liability')
st.markdown("To the extent permitted by applicable law, neither party’s total aggregate liability arising out of or in connection with this Agreement will exceed the price of the Customer’s current subscription."

 

"Neither Party will be liable for indirect or consequential damages arising from or in connection with this Agreement. Consequential damages could include loss of profit, goodwill, data or business reputation. It also could mean the cost of replacing Picanalitika's Service with another."

"Picanalitika provides the Site and the Services “as is” without any warranty or condition of any kind."

"Picanalitika provides a “plug-and-play” tool. A more detailed description of the Services purchased by Customer can be found in the Order Confirmation."

"Picanalitika's Services rely upon Customer’s internet connection, hardware, and network connectivity, and Picanalitika cannot and does not guarantee uninterrupted, secure or error-free operation of the Site."

"Similarly, News Link Search Results are found on the internet; they rely upon the accuracy and timeliness with which online publishers update their articles. Picanalitika cannot and does not make any representation or warranty as to the accuracy, timeliness, quality, or reliability of any information accessed through the Site, especially suitability for a particular business purpose."

"Simply Put: No information obtained from Picanalitika or through the Site, whether oral or written, shall create any warranty not expressly stated in this Agreement. However, nothing herein will reduce warranties, rights, or remedies that Customer is entitled to by law.")

st.header('General')
st.markdown("Governing Law: Picanalitika’s contracting legal entity (appears on the Order Confirmation) determines the area of governing law and jurisdiction for a potential dispute:"

"Picanalitika Legal Entity:"

"Area of Governing Law / Applicable Venue:"

"Indonesia"

"If there were ever a dispute under this Agreement, the Parties will attempt to first resolve it amicably and in good faith."

"Waiver/Severability: Waiving a breach of this Agreement won’t waive any other breach; failing to enforce any right under this Agreement won’t waive any other right. Any provision of this Agreement deemed invalid won’t affect the others, and will assume its closest reasonable interpretation."

"Assignment: Neither party may assign any rights in this Agreement (or the entire Agreement) without the other’s express written consent, unless in connection with its own merger or acquisition."

"Entire Agreement: These Terms along with the Order Confirmation (this Agreement) constitute the entire agreement. Any other spoken conversations, understandings, or terms and conditions, including Purchase Order terms, terms of a tender or RFP, or vendor registration terms are null & void and have no force and effect, even if produced subsequently to signature of the Order Confirmation, and even if Picanalitika has started to provide Services. This Agreement may be only amended in writing if signed by an authorised representative of both Parties."

"Survival: The following provisions shall survive the term of this Agreement: How the Picanalitika License Works, Intellectual Property, Warranty & Liability, Product-Specific Terms, General, and any representations or warranties herein."

"Privacy: Customer agrees to Picanalitika’s privacy policy, available at www.picanalitika.com/Privacy. Notices: Customer agrees that notices, including legal notices, may be sent by electronic mail to the email address listed on the Order Confirmation.")

st.header('Definitions')
st.markdown("Throughout these Terms (and the Order Confirmation), the following terms, when capitalized, will have the following meanings:"

"Agreement: These Terms in conjunction with the Order Confirmation."

"Authorised Use: Use of the Services in strict accordance with the terms of this Agreement."

"Authorised User: Any specific employee or consultant located in the same country as the Customer entity (as defined in the Order Confirmation) that Customer has authorised to use the Services solely on its own behalf. Also defined as Users."

"Customer: Customer or agent of Customer as specified in the Order Confirmation."

"Late Payment Interest: The maximum interest rate allowed by applicable law."

"Picanalitika: The Picanalitika legal entity as specified in the order confirmation. Also defined as ‘its’."

"News Link Search Results: The search result for a specific news article which contains a hyperlink text to the online media source where the article is published."

"Order Confirmation: The statement of work document which defines the Customer, describes Services purchased, contains payment terms & signature blocks."

"The Parties: Customer and Picanalitika, collectively."

"Services: Those services provided by Picanalitika (and third party providers, if any) as described in the Order Confirmation."

"Site: Any site which has the domain or is a subdomain of www.picanalitika.tech. Special Terms: Those terms specified in the ‘Special Terms’ box on the Order Confirmation. If they conflict with any Terms above, the Special Terms override."

"Terms: These terms and conditions of use."

"Third Party Content: Any and all content, advertising, products or other materials belonging to a third party."

"Third Party Sites: Any and all third party websites.")
